refactor(integration): removes uneccessary files and updates help func
