--- .dockerignore
@@ -1,12 +1,0 @@
-lib/
-.git/
-dist/
-node_modules/
-.github/
-*.log
-.huskyrc
-.vscode/
-Docker/
-tmp/
-*.md
-*.txt
