+++ integration/test/http.spec.ts
@@ -42,7 +42,7 @@ describe.each`
         interceptor: { http: parser },
       });
       app = modRef.createNestApplication(adapter);
-      interceptor = modRef.get(getInterceptor(app));
+      interceptor = getInterceptor(app);
       await app.listen(0);
     });
