+++ integration/test/utils/getInterceptor.ts
@@ -1,14 +1,17 @@
 import { INestApplication, INestMicroservice } from '@nestjs/common';
+import { OgmaInterceptor } from '@ogma/nestjs-module';
 
 // just trust me... I hate this...
 export function getInterceptor(
   app: INestApplication | INestMicroservice,
-): string {
-  return Array.from((app as any).container.getModules().values())
-    .filter((module: any) => module.metatype.name === 'OgmaCoreModule')
-    .map((module: any) => {
-      return Array.from<string>(
-        module.providers.keys(),
-      ).filter((prov: string) => prov.includes('APP_INTERCEPTOR'))[0];
-    })[0];
+): OgmaInterceptor {
+  return app.get<OgmaInterceptor>(
+    Array.from((app as any).container.getModules().values())
+      .filter((module: any) => module.metatype.name === 'OgmaCoreModule')
+      .map((module: any) => {
+        return Array.from<string>(
+          module.providers.keys(),
+        ).filter((prov: string) => prov.includes('APP_INTERCEPTOR'))[0];
+      })[0],
+  );
 }
