+++ integration/test/utils/http-promise.ts
@@ -1,4 +1,3 @@
-/* eslint-disable sonarjs/no-identical-functions */
 import { request, RequestOptions } from 'http';
 
 export const httpPromise = (url: string, options?: RequestOptions) =>
@@ -9,26 +8,5 @@ export const httpPromise = (url: string, options?: RequestOptions) =>
       res.on('end', () => resolve(data));
       res.on('error', (err) => reject(err));
     });
-    req.end();
-  });
-
-export const gqlPromise = (url: string, query: object) =>
-  new Promise((resolve, reject) => {
-    const req = request(
-      url,
-      {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      },
-      (res) => {
-        let data = '';
-        res.on('data', (chunk) => (data += chunk));
-        res.on('end', () => resolve(data));
-        res.on('error', (err) => reject(err));
-      },
-    );
-    req.write(Buffer.from(JSON.stringify(query)));
     req.end();
   });
