+++ package.json
@@ -25,10 +25,6 @@
     "build": "lerna run build --stream",
     "check": "npm-check -u",
     "commit": "git-cz",
-    "docker": "docker build",
-    "docker:express": "yarn docker -t ogma/express-test -f Docker/express.Dockerfile .",
-    "docker:fastify": "yarn docker -t ogma/fastify-test -f Docker/fastify.Dockerfile .",
-    "docker:socket": "yarn docker -t ogma/socket-test -f Docker/socket-io.Dockerfile .",
     "format": "prettier --write \"{packages,integration}/**/*.ts\"",
     "lint": "eslint --ext .ts .",
     "setup": "yarn && yarn build",
